GNU99
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2931 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ HTTPServer.m
@@ -287,7 +287,7 @@
     case NSStreamEventHasBytesAvailable:;
         uint8_t buf[16 * 1024];
         uint8_t *buffer = NULL;
-        unsigned int len = 0;
+        NSUInteger len = 0;
         if (![istream getBuffer:&buffer length:&len]) {
             int amount = [istream read:buf maxLength:sizeof(buf)];
             buffer = buf;
